Rename types.ts to types.d.ts
Renamed the types file to a declaration file, which is my personal
preference for typescript type definitions.

type: refactor
scope: types
