+++ src/index.ts
@@ -1,6 +1,6 @@
-export * from './types.js';
 export * from './client.ts';
 export * from './configure.ts';
 export * from './error.ts';
+export type * from './types.d.ts';
 import { Client } from './client.ts';
 export default Client;
